src/main/java/com/sharpon/client/lookup/controller/GenderController.java
package com.sharpon.client.lookup.controller;

import com.sharpon.client.lookup.dto.GenderDto;
import com.sharpon.client.lookup.service.GenderService;
import io.swagger.v3.oas.annotations.tags.Tag;
import lombok.RequiredArgsConstructor;
import lombok.extern.slf4j.Slf4j;
import org.springframework.data.domain.Page;
import org.springframework.data.domain.Pageable;
import org.springframework.data.domain.Sort;
import org.springframework.data.web.PageableDefault;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;

import static com.sharpon.client.lookup.controller.GenderController.BASE_URL;


@Slf4j
@RestController
@RequestMapping(BASE_URL)
@RequiredArgsConstructor
@Tag(name = "Gender API")
public class GenderController {
    public static final String BASE_URL = "/api/v1/gender";
    private final GenderService genderService;

    @PostMapping("/save")
    public GenderDto save(@Validated @RequestBody GenderDto dto) {
        return genderService.save(dto);
    }

    @DeleteMapping("/{id}")
    public ResponseEntity<Void> delete(@PathVariable("id") Long id) {
        genderService.deleteById(id);
        return ResponseEntity.ok().build();
    }

    @GetMapping("/search")
    public ResponseEntity<Page<GenderDto>> search(@PageableDefault(size = 20, sort = {"updatedAt"}, direction = Sort.Direction.ASC) Pageable pageable) {
        return ResponseEntity.ok(genderService.search(pageable));
    }
}
